tests: moved into {integration,unit} dirs
