+++ test/integration/push-jenkins-update.test.js
@@ -7,7 +7,7 @@ const url = require('url')
 const nock = require('nock')
 const supertest = require('supertest')
 
-const app = require('../app')
+const app = require('../../app')
 
 tap.test('Sends POST requests to https://api.github.com/repos/nodejs/node/statuses/<SHA>', (t) => {
   const jenkinsPayload = readFixture('success-payload.json')
@@ -87,6 +87,6 @@ function ignoreQueryParams (pathAndQuery) {
 }
 
 function readFixture (fixtureName) {
-  const content = fs.readFileSync(path.join(__dirname, '_fixtures', fixtureName)).toString()
+  const content = fs.readFileSync(path.join(__dirname, '..', '_fixtures', fixtureName)).toString()
   return JSON.parse(content)
 }
